requirements.txt
###### Requirements without Version Specifiers ######`
tensorflow
matplotlib
numpy
pycopy-colorsys
pillow
opencv-python
scipy
webcolors
